Fix a bug that caused Terminal threads to sleep VERY long under a rare combination of conditions.
If the first transaction of a Terminal happens to be a NEW_ORDER
that under the 1% rule roles back with "Invalid Item", the
transactionEnd timestamp was not set (left uninitialized). If
That happened at a time where the following (buggy) calculation
of sleep time for limitTxnsPerMin happens to be < -2^31, the
casting to int caused the number of sleep millis to be 2 billion
and some. The thread would never check the end flag or process
any more transactions for a long time.

Conflicts:
	src/client/jTPCCTerminal.java