Adding a flag to the props files that can be used enable/disable the fixed Warehouse association for a Terminal. The default is enabled, which is the backwards compatible behaviour of 4.x.
BenchmarkSQL currently does not implement Terminals in the way,
TPC-C describes them. Instead it implements a combination of Terminal
(the part generating transaction input data) and what should be
an application thread (the part processing the input data over
a database connection). In this implementation the fixed Warehouse
per Terminal translates to a fixed Warehouse per database connection.

According to the TPC-C specs there should be 10 Terminals per Warehouse.
In the current implementation, that is not practical because at a
perfectly reasonable scaling factor of 1,000 Warehouses (building an
initial database of approx. 100GB), that would require 10,000
database connections.

This discrepancy leads to a distribution of database access that is
different from what TPC-C defines since most transactions focus on
the data, associated with the Terminal's home Warehouse.

Setting the new config option terminalWarehouseFixed=false will cause
the Terminal to generate a random W_ID for each new transaction and
thus distribute the database access evenly over all configured
Warehouses.