Major code overhaul. Moved all transaction specific terminal data generation and processing into new module jTPCCTData. Moved JDBC connection and all prepared statements into new module jTPCCConnection. These steps are in preparation for sharing this functionality with a new benchmark driver, that will measure end-user terminal response times as defined by TPC-C.
The transaction profiles have been redone based on a more
narrow interpretation of the TPC-C specs. Earlier changes
to the 5.0 developement tree already made performance results
of pre-5.0 not compare to 5.0. Namely the addition of foreign
keys and enabling the 60% lookup of customer by last name in
the PAYMENT and ORDER_STATUS transactions. So there is no
point in limiting the number of performance result altering
changes.

There still is a very bad anomaly with the Firebird port
in what appears like FB is ignoring the "... FOR UPDATE" part
of a SELECT statement and will happily return the same row
to two or more concurrent transactions. This leads to
PRIMARY KEY violations in the NEW_ORDER transaction. But it
should not be a show stopper at this point.