Change the transaction weight variables to double to be able to create a transaction mix that is closer to the 23 Card Deck implementation, suggested in the TPC-C Spec.
In version 5.0 and prior, there is an imbalance between
NEW_ORDER and DELIVERY. On average, more new orders are
created than being delivered. This causes the new_order
table and the active portions of oorder and order_line to
constantly grow. This way the benchmark never has a chance
to reach a steady state.